refactor: removed formInfoContent and replaced with form provider
