+++ shesha-reactjs/src/providers/form/contexts.tsx
@@ -1,5 +1,5 @@
 import { FormInstance } from 'antd';
-import { createContext } from 'react';
+import { ReactNode, createContext } from 'react';
 import { IFormValidationErrors, IToolboxComponent, IToolboxComponentGroup } from '../../interfaces';
 import {
   DEFAULT_FORM_SETTINGS,
@@ -11,6 +11,7 @@ import {
   IFlatComponentsStructure,
   IFormAction,
   IFormActions,
+  IFormDesignerActionFlag,
   IFormSection,
   IFormSettings,
   ROOT_COMPONENT_KEY,
@@ -42,6 +43,15 @@ export interface IFormStateInternalContext {
    * If true, indicates that list of visible components are calculated
    */
   visibleComponentIdsIsSet: boolean;
+  /**
+   * Action flag to determine event which was triggered
+   */
+  actionFlag?: { [key in IFormDesignerActionFlag]?: boolean };
+
+  /**
+   * External components rendered on the designer mode
+   */
+  renderToolbarRightButtons?: ReactNode[];
 }
 
 export interface IFormStateContext extends IFormStateInternalContext, IFlatComponentsStructure {}
@@ -99,6 +109,12 @@ export interface IFormActionsContext {
   getSection: (id: string, name: string) => FormSection;
 
   getToolboxComponent: (type: string) => IToolboxComponent;
+
+  /** Flag set to determine action triggered from external source */
+  setActionFlag: (flag: IFormDesignerActionFlag) => void;
+
+  /** Set form designer toolbar button */
+  setToolbarRightButton: (button: ReactNode) => void;
 }
 
 /** Form initial state */
@@ -114,6 +130,8 @@ export const FORM_CONTEXT_INITIAL_STATE: IFormStateContext = {
   sections: [],
   context: null,
   formSettings: DEFAULT_FORM_SETTINGS,
+  actionFlag: {},
+  renderToolbarRightButtons: [],
 };
 
 export interface ConfigurableFormInstance extends IFormActionsContext, IFormStateContext {}
