+++ shesha-reactjs/src/providers/form/models.ts
@@ -21,6 +21,8 @@ export type VisibilityType = 'Yes' | 'No' | 'Removed';
 
 export type PropertySettingMode = 'value' | 'code';
 
+export type IFormDesignerActionFlag = 'done' | 'redo' | 'publish' | 'settings' | 'undo' | 'version';
+
 export interface IPropertySetting {
   _mode?: PropertySettingMode;
   _value?: any;
