--- shesha-reactjs/src/providers/formInfoContent/models.ts
@@ -1,6 +1,0 @@
-export type IFormDesignerActionFlag = 'done' | 'redo' | 'publish' | 'settings' | 'undo' | 'version';
-
-export enum ActionFlag {
-  name = 'ActionFlag',
-  render = 'RenderToolbarRightButtons',
-}
